IOS-7688: Improve project file structure
