--- Tangem/Common/UI/AuxiliaryTypes.swift
@@ -1,23 +1,0 @@
-//
-//  AuxiliaryTypes.swift
-//  Tangem
-//
-//  Created by skibinalexander on 16.07.2024.
-//  Copyright © 2024 Tangem AG. All rights reserved.
-//
-
-import SwiftUI
-
-@available(iOS 16.0, *)
-enum RoundedCornersVerticalEdge {
-    case topEdge
-    case bottomEdge
-    case all
-}
-
-struct RoundedCornersConfiguration {
-    var topLeadingRadius: CGFloat = 0.0
-    var bottomLeadingRadius: CGFloat = 0.0
-    var bottomTrailingRadius: CGFloat = 0.0
-    var topTrailingRadius: CGFloat = 0.0
-}
